Start separating IFC conversion methods.
