+++ src/Elements/Serialization/IFC/IFCExtensions.cs
@@ -7,14 +7,13 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 
 namespace Elements.Serialization.IFC
 {
     /// <summary>
     /// Extension methods for converting IFC types to Element types.
     /// </summary>
-    public static class IFCExtensions
+    public static partial class IFCExtensions
     {
         /// <summary>
         /// Load a model from IFC.
@@ -142,36 +141,31 @@ namespace Elements.Serialization.IFC
                     if(e is StandardWall)
                     {
                         var w = (StandardWall)e;
-                        var ifcWall = w.ToIfcWallStandardCase(context, ifc);
-                        products.Add(ifcWall);
+                        w.ToIfcWallStandardCase(context, ifc, products);
                     }
 
                     if(e is Wall)
                     {
                         var w = (Wall)e;
-                        var ifcWall = w.ToIfcWall(context, ifc);
-                        products.Add(ifcWall);
+                        w.ToIfcWall(context, ifc, products);
                     }
 
                     if(e is Beam)
                     {
                         var b = (Beam)e;
-                        var ifcBeam = b.ToIfcBeam(context, ifc);
-                        products.Add(ifcBeam);
+                        b.ToIfcBeam(context, ifc, products);
                     }
 
                     if(e is Floor)
                     {
                         var f = (Floor)e;
-                        var ifcSlab = f.ToIfcSlab(context, ifc);
-                        products.Add(ifcSlab);
+                        f.ToIfcSlab(context, ifc, products);
                     }
 
                     if(e is Space)
                     {
                         var s = (Space)e;
-                        var ifcSpace = s.ToIfcSpace(context, ifc);
-                        products.Add(ifcSpace);
+                        s.ToIfcSpace(context, ifc, products);
                     }
                 }
                 catch(Exception ex)
@@ -315,118 +309,7 @@ namespace Elements.Serialization.IFC
             // IFC materials is a fucking trainwreck. Just return a default.
             return new Material(material.Name, Colors.Gray, 0.0f, 0.0f);
         }
-
-        /// <summary>
-        /// Convert a wall to an IfcWallStandardCase
-        /// </summary>
-        /// <param name="wall"></param>
-        /// <param name="context"></param>
-        /// <param name="doc"></param>
-        /// <returns></returns>
-        private static IfcWallStandardCase ToIfcWallStandardCase(this StandardWall wall, IfcRepresentationContext context, Document doc)
-        {
-            var sweptArea = wall.CenterLine.Thicken(wall.Thickness()).ToIfcArbitraryClosedProfileDef(doc);
-            var extrudeDirection = Vector3.ZAxis.ToIfcDirection();
-
-            // We don't use the Wall's transform for positioning, because
-            // our walls have a transform that lays the wall "flat". Just
-            // use a identity transform or a transform that includes
-            // the elevation.
-            var position = new Transform().ToIfcAxis2Placement3D(doc);
-            var repItem = new IfcExtrudedAreaSolid(sweptArea, position, 
-                extrudeDirection, new IfcPositiveLengthMeasure(wall.Height));
-            var rep = new IfcShapeRepresentation(context, "Body", "SweptSolid", new List<IfcRepresentationItem>{repItem});
-            var productRep = new IfcProductDefinitionShape(new List<IfcRepresentation>{rep});
-            var id = IfcGuid.ToIfcGuid(Guid.NewGuid());
-            var localPlacement = new Transform(0,0,wall.CenterLine.Start.Z).ToIfcLocalPlacement(doc);
-            var ifcWall = new IfcWallStandardCase(new IfcGloballyUniqueId(id), 
-                null, wall.Name, null, null, localPlacement, productRep, null);
-
-            doc.AddEntity(sweptArea);
-            doc.AddEntity(extrudeDirection);
-            doc.AddEntity(position);
-            doc.AddEntity(repItem);
-            doc.AddEntity(rep);
-            doc.AddEntity(localPlacement);
-            doc.AddEntity(productRep);
-            doc.AddEntity(ifcWall);
-
-            return ifcWall;
-        }
-
-        private static IfcWall ToIfcWall(this Wall wall, IfcRepresentationContext context, Document doc)
-        {
-            var sweptArea = wall.Transform.OfProfile(wall.Profile).Perimeter.ToIfcArbitraryClosedProfileDef(doc);
-            var extrudeDirection = Vector3.ZAxis.ToIfcDirection();
-
-            // We don't use the Wall's transform for positioning, because
-            // our walls have a transform that lays the wall "flat". Just
-            // use a identity transform or a transform that includes
-            // the elevation.
-            var position = new Transform().ToIfcAxis2Placement3D(doc);
-            var repItem = new IfcExtrudedAreaSolid(sweptArea, position, 
-                extrudeDirection, new IfcPositiveLengthMeasure(wall.Height));
-            var rep = new IfcShapeRepresentation(context, "Body", "SweptSolid", new List<IfcRepresentationItem>{repItem});
-            var productRep = new IfcProductDefinitionShape(new List<IfcRepresentation>{rep});
-            var id = IfcGuid.ToIfcGuid(Guid.NewGuid());
-            var localPlacement = new Transform().ToIfcLocalPlacement(doc);
-            var ifcWall = new IfcWall(new IfcGloballyUniqueId(id), 
-                null, wall.Name, null, null, localPlacement, productRep, null);
-
-            doc.AddEntity(sweptArea);
-            doc.AddEntity(extrudeDirection);
-            doc.AddEntity(position);
-            doc.AddEntity(repItem);
-            doc.AddEntity(rep);
-            doc.AddEntity(localPlacement);
-            doc.AddEntity(productRep);
-            doc.AddEntity(ifcWall);
-
-            return ifcWall;
-        }
-
-        /// <summary>
-        /// Convert a beam to an IfcBeam
-        /// </summary>
-        /// <param name="beam"></param>
-        /// <param name="context"></param>
-        /// <param name="doc"></param>
-        /// <returns></returns>
-        private static IfcBeam ToIfcBeam(this Beam beam, IfcRepresentationContext context, Document doc)
-        {
-            var sweptArea = beam.ElementType.Profile.Perimeter.ToIfcArbitraryClosedProfileDef(doc);
-            var line = beam.Curve as Line;
-            if(line == null) {
-                throw new Exception("The beam could not be exported to IFC. Only linear beams are currently supported.");
-            }
-            
-            // We use the Z extrude direction because the direction is 
-            // relative to the local placement, which is a transform at the
-            // beam's end with the Z axis pointing along the direction.
-            
-            var extrudeDirection = Vector3.ZAxis.ToIfcDirection();
-
-            var position = new Transform().ToIfcAxis2Placement3D(doc);
-            var repItem = new IfcExtrudedAreaSolid(sweptArea, position, 
-                extrudeDirection, new IfcPositiveLengthMeasure(beam.Curve.Length()));
-            var localPlacement = beam.Curve.TransformAt(0.0).ToIfcLocalPlacement(doc);
-            // var placement = beam.Transform.ToIfcAxis2Placement3D(doc);
-            var rep = new IfcShapeRepresentation(context, "Body", "SweptSolid", new List<IfcRepresentationItem>{repItem});
-            var productRep = new IfcProductDefinitionShape(new List<IfcRepresentation>{rep});
-            var ifcBeam = new IfcBeam(IfcGuid.ToIfcGuid(Guid.NewGuid()), null, null, null, null, localPlacement, productRep, null);
-            
-            doc.AddEntity(sweptArea);
-            doc.AddEntity(extrudeDirection);
-            doc.AddEntity(position);
-            doc.AddEntity(repItem);
-            doc.AddEntity(rep);
-            doc.AddEntity(localPlacement);
-            doc.AddEntity(productRep);
-            doc.AddEntity(ifcBeam);
-
-            return ifcBeam;
-        }
-
+        
         /// <summary>
         /// Convert a polygon to an IfcArbitraryClosedProfileDef.
         /// </summary>
@@ -513,217 +396,6 @@ namespace Elements.Serialization.IFC
         }
 
         /// <summary>
-        /// Convert an IfcSlab to a Floor.
-        /// </summary>
-        /// <param name="slab">An IfcSlab.</param>
-        /// <param name="openings"></param>
-        private static Floor ToFloor(this IfcSlab slab, IEnumerable<IfcOpeningElement> openings)
-        {
-            var transform = new Transform();
-            transform.Concatenate(slab.ObjectPlacement.ToTransform());
-
-            // Check if the slab is contained in a building storey
-            foreach (var cis in slab.ContainedInStructure)
-            {
-                transform.Concatenate(cis.RelatingStructure.ObjectPlacement.ToTransform());
-            }
-
-            var repItems = slab.Representation.Representations.SelectMany(r => r.Items);
-            if (!repItems.Any())
-            {
-                throw new Exception("The provided IfcSlab does not have any representations.");
-            }
-
-            var solid = slab.RepresentationsOfType<IfcExtrudedAreaSolid>().FirstOrDefault();
-            if (solid == null)
-            {
-                return null;
-                // throw new Exception("No IfcExtrudedAreaSolid could be found in the provided IfcSlab.");
-            }
-
-            var floorType = new FloorType($"{Guid.NewGuid().ToString()}_floor_type", new List<MaterialLayer>{new MaterialLayer(new Material("slab",Colors.Green), (IfcLengthMeasure)solid.Depth)});
-            var outline = (Polygon)solid.SweptArea.ToICurve();
-            var solidTransform = solid.Position.ToTransform();
-            var floor = new Floor(new Profile(outline), solidTransform, solid.ExtrudedDirection.ToVector3(), 
-                floorType, 0, transform);
-            floor.Name = slab.Name;
-            
-            return floor;
-        }
-
-        /// <summary>
-        /// Convert and IfcSpace to a Space.
-        /// </summary>
-        /// <param name="space">An IfcSpace.</param>
-        /// <returns></returns>
-        private static Space ToSpace(this IfcSpace space)
-        {
-            var transform = new Transform();
-
-            var repItems = space.Representation.Representations.SelectMany(r => r.Items);
-            if (!repItems.Any())
-            {
-                throw new Exception("The provided IfcSlab does not have any representations.");
-            }
-
-            var localPlacement = space.ObjectPlacement.ToTransform();
-            transform.Concatenate(localPlacement);
-            
-            var foundSolid = repItems.First();
-            var material = new Material("space", new Color(1.0f, 0.0f, 1.0f, 0.5f), 0.0f, 0.0f);
-            if (foundSolid.GetType() == typeof(IfcExtrudedAreaSolid))
-            {
-                var solid = (IfcExtrudedAreaSolid)foundSolid;
-                var profileDef = (IfcArbitraryClosedProfileDef)solid.SweptArea;
-                transform.Concatenate(solid.Position.ToTransform());
-                var pline = (IfcPolyline)profileDef.OuterCurve;
-                var outline = pline.ToPolygon(true);
-                var result = new Space(new Profile(outline), (IfcLengthMeasure)solid.Depth, 0.0, material, transform);
-                result.Name = space.Name;
-                return result;
-            }
-            else if (foundSolid.GetType() == typeof(IfcFacetedBrep))
-            {
-                var solid = (IfcFacetedBrep)foundSolid;
-                var shell = solid.Outer;
-                var newSolid = new Solid();
-                for(var i=0; i< shell.CfsFaces.Count; i++)
-                {
-                    var f = shell.CfsFaces[i];
-                    foreach (var b in f.Bounds)
-                    {
-                        var loop = (IfcPolyLoop)b.Bound;
-                        var poly = loop.Polygon.ToPolygon();
-                        newSolid.AddFace(poly);
-                    }
-                }
-                var result = new Space(newSolid, transform);
-
-                result.Name = space.Name;
-                return result;
-            }
-
-            return null;
-        }
-
-        /// <summary>
-        /// Convert an IfcWallStandardCase to a Wall.
-        /// </summary>
-        /// <param name="wall">An IfcWallStandardCase.</param>
-        /// <param name="openings">A collection of IfcOpeningElement belonging to this wall.</param>
-        /// <param name="wallType">The wall's wall type.</param>
-        private static Wall ToWall(this IfcWallStandardCase wall, 
-            IEnumerable<IfcOpeningElement> openings, WallType wallType)
-        {
-            var transform = new Transform();
-            transform.Concatenate(wall.ObjectPlacement.ToTransform());
-
-            // An extruded face solid.
-            var solid = wall.RepresentationsOfType<IfcExtrudedAreaSolid>().FirstOrDefault();
-            if(solid == null)
-            {
-                // It's possible that the rep is a boolean.
-                var boolean = wall.RepresentationsOfType<IfcBooleanClippingResult>().FirstOrDefault();
-                if(boolean != null)
-                {
-                    solid = boolean.FirstOperand.Choice as IfcExtrudedAreaSolid;
-                    if(solid == null)
-                    {
-                        solid = boolean.SecondOperand.Choice as IfcExtrudedAreaSolid; 
-                    }
-                }
-
-                // if(solid == null)
-                // {
-                //     throw new Exception("No usable solid was found when converting an IfcWallStandardCase to a Wall.");
-                // }
-            }
-            
-            // A centerline wall with material layers.
-            // var axis = (Polyline)wall.RepresentationsOfType<IfcPolyline>().FirstOrDefault().ToICurve(false);
-            
-            foreach (var cis in wall.ContainedInStructure)
-            {
-                cis.RelatingStructure.ObjectPlacement.ToTransform().Concatenate(transform);
-            }
-
-            // var os = openings.Select(o=>o.ToOpening()).ToArray();
-
-            if(solid != null)
-            {
-                var c = solid.SweptArea.ToICurve();
-                if(c is Polygon)
-                {
-                    transform.Concatenate(solid.Position.ToTransform());
-                    var result = new Wall((Polygon)c, wallType, (IfcLengthMeasure)solid.Depth, transform);
-
-                    result.Name = wall.Name;
-                    return result;
-                }
-            }
-            return null;
-        }
-
-        /// <summary>
-        /// Convert an IfcBeam to a beam.
-        /// </summary>
-        /// <param name="beam"></param>
-        private static Beam ToBeam(this IfcBeam beam)
-        {
-            var elementTransform = beam.ObjectPlacement.ToTransform();
-            
-            var solid = beam.RepresentationsOfType<IfcExtrudedAreaSolid>().FirstOrDefault();
-
-            // foreach (var cis in beam.ContainedInStructure)
-            // {
-            //     cis.RelatingStructure.ObjectPlacement.ToTransform().Concatenate(transform);
-            // }
-
-            if(solid != null)
-            {
-                var solidTransform = solid.Position.ToTransform();
-
-                var c = solid.SweptArea.ToICurve();
-                if(c is Polygon)
-                {
-                    var cl = new Line(Vector3.Origin, 
-                        solid.ExtrudedDirection.ToVector3(), (IfcLengthMeasure)solid.Depth);
-                    var framingType = new StructuralFramingType(Guid.NewGuid().ToString(), new Profile((Polygon)c), BuiltInMaterials.Steel);
-                    var result = new Beam(solidTransform.OfLine(cl), framingType, 0.0, 0.0, elementTransform);
-                    result.Name = beam.Name;
-                    return result; 
-                }
-            }
-            return null;
-        }
-
-        /// <summary>
-        /// Convert an IfcColumn to a Column.
-        /// </summary>
-        /// <param name="column"></param>
-        private static Column ToColumn(this IfcColumn column)
-        {
-            var elementTransform = column.ObjectPlacement.ToTransform();
-            
-            var solid = column.RepresentationsOfType<IfcExtrudedAreaSolid>().FirstOrDefault();
-            foreach (var cis in column.ContainedInStructure)
-            {
-                cis.RelatingStructure.ObjectPlacement.ToTransform().Concatenate(elementTransform);
-            }
-
-            if(solid != null)
-            {
-                var solidTransform = solid.Position.ToTransform();
-                var c = solid.SweptArea.ToICurve();
-                var framingType = new StructuralFramingType(Guid.NewGuid().ToString(), new Profile((Polygon)c), BuiltInMaterials.Steel);
-                var result = new Column(solidTransform.Origin, (IfcLengthMeasure)solid.Depth, framingType, elementTransform, 0.0, 0.0);
-                result.Name = column.Name;
-                return result;
-            }
-            return null;
-        }
-
-        /// <summary>
         /// Convert an IfcOpening to an Opening.
         /// </summary>
         /// <param name="opening"></param>
@@ -744,74 +416,6 @@ namespace Elements.Serialization.IFC
                 return newOpening;
             }
             return null;
-        }
-
-        /// <summary>
-        /// Convert a Floor to an IfcSlab.
-        /// </summary>
-        /// <param name="floor"></param>
-        /// <param name="context"></param>
-        /// /// <param name="doc"></param>
-        private static IfcSlab ToIfcSlab(this Floor floor, IfcRepresentationContext context, Document doc)
-        {
-            var position = floor.Transform.ToIfcAxis2Placement3D(doc);
-            var sweptArea = floor.Profile.Perimeter.ToIfcArbitraryClosedProfileDef(doc);
-            var extrudeDirection = Vector3.ZAxis.ToIfcDirection();
-            var repItem = new IfcExtrudedAreaSolid(sweptArea, position, 
-                extrudeDirection, new IfcPositiveLengthMeasure(floor.Thickness()));
-            var localPlacement = new Transform().ToIfcLocalPlacement(doc);
-
-            var placement = floor.Transform.ToIfcAxis2Placement3D(doc);
-            var rep = new IfcShapeRepresentation(context, "Body", "SweptSolid", new List<IfcRepresentationItem>{repItem});
-            var productRep = new IfcProductDefinitionShape(new List<IfcRepresentation>{rep});
-
-            var slab = new IfcSlab(IfcGuid.ToIfcGuid(Guid.NewGuid()), null, null, null, 
-                null, localPlacement, productRep, null, IfcSlabTypeEnum.FLOOR);
-
-            doc.AddEntity(sweptArea);
-            doc.AddEntity(extrudeDirection);
-            doc.AddEntity(position);
-            doc.AddEntity(repItem);
-            doc.AddEntity(rep);
-            doc.AddEntity(localPlacement);
-            doc.AddEntity(productRep);
-            doc.AddEntity(slab);
-
-            foreach(var o in floor.Openings)
-            {
-                var opening = o.ToIfcOpeningElement(context, doc);
-                var voidRel = new IfcRelVoidsElement(IfcGuid.ToIfcGuid(Guid.NewGuid()), null, slab, opening);
-                slab.HasOpenings.Add(voidRel);
-                doc.AddEntity(voidRel);
-            }
-
-            return slab;
-        }
-
-        private static IfcSpace ToIfcSpace(this Space space, IfcRepresentationContext context, Document doc)
-        {
-            var position = space.Transform.ToIfcAxis2Placement3D(doc);
-            var sweptArea = space.Profile.Perimeter.ToIfcArbitraryClosedProfileDef(doc);
-            var extrudeDirection = Vector3.ZAxis.ToIfcDirection();
-            var repItem = new IfcExtrudedAreaSolid(sweptArea, position, 
-                extrudeDirection, new IfcPositiveLengthMeasure(space.ExtrudeDepth));
-            var localPlacement = new Transform().ToIfcLocalPlacement(doc);
-            var rep = new IfcShapeRepresentation(context, "Body", "SweptSolid", new List<IfcRepresentationItem>{repItem});
-            var productRep = new IfcProductDefinitionShape(new List<IfcRepresentation>{rep});
-
-            var ifcSpace = new IfcSpace(IfcGuid.ToIfcGuid(Guid.NewGuid()), null, null, null, 
-                null, localPlacement, productRep, null, IfcElementCompositionEnum.ELEMENT, IfcInternalOrExternalEnum.NOTDEFINED, new IfcLengthMeasure(space.Transform.Origin.Z));
-            
-            doc.AddEntity(sweptArea);
-            doc.AddEntity(extrudeDirection);
-            doc.AddEntity(position);
-            doc.AddEntity(repItem);
-            doc.AddEntity(rep);
-            doc.AddEntity(localPlacement);
-            doc.AddEntity(productRep);
-            doc.AddEntity(ifcSpace);
-
-            return ifcSpace;
         }
 
         private static Solid Representations(this IfcProduct product)
